Also move out the global stuff from view.cpp/h to root.cpp/h
