+++ UI/ChatScreen.cpp
@@ -1,4 +1,5 @@
 #include "ppsspp_config.h"
+#include "ui/root.h"
 #include "ui/ui_context.h"
 #include "ui/view.h"
 #include "ui/viewgroup.h"
