+++ ext/native/ui/screen.cpp
@@ -2,6 +2,7 @@
 #include "base/logging.h"
 #include "base/timeutil.h"
 #include "input/input_state.h"
+#include "ui/root.h"
 #include "ui/screen.h"
 #include "ui/ui.h"
 #include "ui/view.h"
